Bring back the resizing and aspect ratio support
The onscreen window can now be resized and the code to track the
aspect ratio has been copied back in from the original
cogl-basic-video-player example. The effect is now passed the
rectangle to paint in and an orthographic projection is set up. The
point sprite effect is now set up to use the position attribute and
gl_PointCoord rather than gl_FragCoord to work out where in the video
the fragment is.