Add a no-op effect
This adds a no-op effect which just directly draws the sink's pipeline
like the basic video player would. The code to draw the borders is
kept in a separate file so that it could be used by other effects as
well. It's not currently possible to use the effect because the player
always just chooses the first effect.